0_pytorch_integrated/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
tqdm>=4.60.0
flask>=2.0.0
pillow>=8.0.0
torchtext>=0.11.0
umap-learn>=0.5.0
seaborn>=0.11.0
pandas>=1.3.0
